Iterators Refactoring (#9)
