+++ src/replicate.ts
@@ -1,16 +1,18 @@
 import { InputTypeError, IllegalArgumentError } from './commons/ErrorModels';
-import { RepeatIterator } from './repeat';
-import { CloneableIterator, PrimitiveIterator } from './base';
+import {
+    CloneableIterator,
+    PrimitiveIterator,
+    RepeatIterator,
+} from './commons/Iterators';
 
 function replicate(count: number, item: number | string) {
     if (
         typeof count !== 'number' ||
-        (typeof item !== 'number' &&
-            typeof item !== 'string')
+        (typeof item !== 'number' && typeof item !== 'string')
     ) {
         throw new InputTypeError();
     }
-    
+
     if (count < 1) {
         throw new IllegalArgumentError('replicate count must be larger than 1');
     }
